+++ Pod/Classes/URBNRegexRule.swift
@@ -1,0 +1,30 @@
+//
+//  URBNRegexRule.swift
+//  URBNValidator
+//
+//  Created by Joseph Ridenour on 12/22/15.
+//
+//
+
+import Foundation
+
+
+public class URBNRegexRule: URBNBaseRule, URBNRequirement {
+    internal var pattern: String
+    public var isRequired: Bool = false
+    
+    public init(pattern: String, localizationKey: String? = nil) {
+        self.pattern = pattern
+        super.init(localizationKey: localizationKey)
+    }
+    
+    public override func validateValue<T: AnyObject>(value: T?) -> Bool {
+        if !isRequired && value == nil { return true }
+        
+        let pred = NSPredicate(format: "SELF MATCHES[cd] %@", self.pattern)
+        
+        return pred.evaluateWithObject(value)
+    }
+    
+    public static let emailPattern = "^(?:(?:(?:(?:[a-zA-Z0-9_!#\\$\\%&'*+/=?\\^`{}~|\\-]+)(?:\\.(?:[a-zA-Z0-9_!#\\$\\%&'*+/=?\\^`{}~|\\-]+))*)|(?:\"(?:\\\\[^\\r\\n]|[^\\\\\"])*\")))\\@(?:(?:(?:(?:[a-zA-Z0-9_!#\\$\\%&'*+/=?\\^`{}~|\\-]+)(?:\\.(?:[a-zA-Z0-9_!#\\$%&'*+/=?\\^`{}~|\\-]+))*)|(?:\\[(?:\\\\\\S|[\\x21-\\x5a\\x5e-\\x7e])*\\])))$"
+}
